@deprecate Remove init command
